videostore/requirements.txt
appdirs==1.4.3
certifi==2018.4.16
chardet==3.0.4
Django==1.8
django-filter==1.1.0
djangorestframework==3.1
djangorestframework-jwt==1.11.0
djutils==0.3.2
gunicorn==19.7.1
idna==2.6
ipaddress==1.0.22
packaging==17.1
Pillow==5.1.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
requests==2.18.4
urllib3==1.21.1
django-simple-history==1.9.1